Improve init.sh

+++ init.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 set -e
 git submodule update --init --recursive
-ln -s ${HOME}/.vim/vimrc ${HOME}/.vimrc
-vim -c 'BundleInstall' -c 'quit'
+if [ ! -e ${HOME}/.vimrc ]; then
+  ln -s ${HOME}/.vim/vimrc ${HOME}/.vimrc
+fi
+vim -c 'BundleInstall'
+./bundle/YouCompleteMe/install.sh --clang-completer
